+++ frontend/src/components/ui/table.tsx
@@ -1,8 +1,12 @@
 "use client"
 
 import * as React from "react"
-
+import { Bolt } from "lucide-react"
+import { FaGithub } from "react-icons/fa"
 import { cn } from "@/lib/utils"
+import { useState } from "react"
+import { Switch } from "./switch"
+import { GithubRepo } from "@/types/githube_types"
 
 function Table({ className, ...props }: React.ComponentProps<"table">) {
   return (
@@ -78,11 +82,51 @@ function TableCell({ className, ...props }: React.ComponentProps<"td">) {
   )
 }
 
+const RepoRow = ({ repo }: { repo: GithubRepo }) => {
+  const [isActive, setIsActive] = useState(repo.active);
+
+  const handleToggle = () => {
+      setIsActive(prevState => !prevState);
+  };
+
+  return (
+      <TableRow>
+          <TableCell className="font-medium">
+              <div className="flex items-center gap-4">
+                  <span>{repo.full_name.split("/")[1]}</span>
+                  <span className={`py-0.5 px-2.5 text-xs font-semibold border rounded-full ${repo.private ? "bg-gray-100 text-gray-800 border-gray-300" : "bg-green-100 text-green-800 border-green-300"}`}>
+                      {repo.private ? "Private" : "Public"}
+                  </span>
+              </div>
+          </TableCell>
+          <TableCell>
+              <a href={repo.html_url} target="_blank" rel="noopener noreferrer" className="text-neutral-500 hover:text-neutral-900 transition-colors">
+                  <FaGithub size={20} />
+              </a>
+          </TableCell>
+          <TableCell>
+              <button className="text-neutral-500 hover:text-neutral-900 transition-colors" aria-label="Settings">
+                  <Bolt size={20} />
+              </button>
+          </TableCell>
+          <TableCell>
+              <Switch
+                  id={`active-${repo.id}`}
+                  checked={isActive}
+                  onCheckedChange={handleToggle}
+                  aria-label={`Activate ${repo.full_name}`}
+              />
+          </TableCell>
+      </TableRow>
+  );
+};
+
 export {
   Table,
   TableHeader,
   TableBody,
   TableHead,
   TableRow,
-  TableCell
+  TableCell,
+  RepoRow
 }
