+++ frontend/src/types/githube_types.ts
@@ -1,0 +1,7 @@
+export interface GithubRepo {
+    id: number;
+    full_name: string;
+    private: boolean;
+    html_url: string;
+    active: boolean;
+}
